Prompt store fixes surfaced by E2E
Two defects the first E2E revealed, both fixed against a fresh local
apply-and-publish loop:

1. CHECK constraint too tight. Original required
   `approved_by IS NOT NULL AND approved_at IS NOT NULL` when active=true.
   `vani_prompt.approved_by` references `vani_user(id)`, but the JWT
   carries a `vn_users.id` — same bridge gap already called out on
   `vara_jd.created_by`. Loosened to `approved_at IS NOT NULL` so an
   override can go live today; the constraint tightens back to include
   `approved_by` in a later migration when the vn_user → vani_user
   bridge lands. Documented inline.

2. `jsonb_build_object($key, …)` needs `$key::text` — CLAUDE.md lesson 2.
   The audit_log insert for override_activated was missing an explicit
   cast on the version param ($5), so pg couldn't infer its type and
   the whole insert threw `could not determine data type of parameter $5`.
   Cast added.

E2E now 16/16 against local backend + local Postgres:
  list → detail → variable validation (Save disabled when a required
  {{var}} is dropped) → save v1 → resolver returns override → revert
  → history preserved → resolver falls back to system → re-save v2.

+++ backend/src/vara/vara.routes.ts
@@ -929,7 +929,7 @@ export function createVaraRouter(pool: Pool): Router {
       await client.query(
         `INSERT INTO vani_audit_log (tenant_id, agent_id, actor_type, actor_id, entity, entity_id, action, before, after)
          SELECT $1, a.id, 'human', $2, 'vani_prompt', $3, 'override_activated',
-                '{}'::jsonb, jsonb_build_object('key', $4::text, 'version', $5)
+                '{}'::jsonb, jsonb_build_object('key', $4::text, 'version', $5::int)
            FROM vani_agent a WHERE a.code = 'vara'`,
         [vani.id, auth.user_id, ins.rows[0].id, key, nextVersion],
       );
